requirements.txt
pandas==2.0.3
numpy==1.24.3
shiny==1.2.1
shinyswatch==0.8.0
duckdb==1.1.1
finnhub-python==2.4.22
yfinance==0.2.51
matplotlib==3.7.2